requirements.txt
selenium
undetected-chromedriver
seleniumbase
google-generativeai
requests
